+++ extingress/haproxy_delay_traffic_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/steadybit/action-kit/go/action_kit_api/v2"
 	"github.com/steadybit/extension-kit/extutil"
 	"github.com/steadybit/extension-kubernetes/v2/client"
+	"github.com/steadybit/extension-kubernetes/v2/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	networkingv1 "k8s.io/api/networking/v1"
@@ -322,5 +323,6 @@ func TestHAProxyDelayTrafficAction_Prepare(t *testing.T) {
 }
 
 func getTestClient(stopCh <-chan struct{}, objects ...runtime.Object) *client.Client {
-	return client.CreateClient(testclient.NewClientset(objects...), stopCh, "", client.MockAllPermitted())
+	dynamicClient := testutil.NewFakeDynamicClient()
+	return client.CreateClient(testclient.NewClientset(objects...), stopCh, "", client.MockAllPermitted(), dynamicClient)
 }