+++ extnode/node_discovery_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/steadybit/extension-kubernetes/v2/client"
 	"github.com/steadybit/extension-kubernetes/v2/extconfig"
+	"github.com/steadybit/extension-kubernetes/v2/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
@@ -153,5 +154,6 @@ func Test_nodeDiscovery(t *testing.T) {
 }
 
 func getTestClient(stopCh <-chan struct{}, objects ...runtime.Object) *client.Client {
-	return client.CreateClient(testclient.NewClientset(objects...), stopCh, "", client.MockAllPermitted())
+	dynamicClient := testutil.NewFakeDynamicClient()
+	return client.CreateClient(testclient.NewClientset(objects...), stopCh, "", client.MockAllPermitted(), dynamicClient)
 }