+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Changelog
 
+
 ## (next version)
 
 - Advice support new service validation step to ease experiment creation
+
+## v2.6.18
+
+- Update dependencies
+- Added categories to advice (for future use in the UI)
+
+## v2.6.17
+
+- Update dependencies
+- Added a discover for Argo Rollouts and added a rollout restart attack
+
+## v2.6.16
+
+- Copy the pod attributes by ownership to statefulset/daemonsets/deployments/replicasets instead of the selector
 
 ## v2.6.15